MOV move spectra_embedding from decomposition to manifold
