+++ sklearn/decomposition/__init__.py
@@ -12,7 +12,6 @@ from .fastica_ import FastICA, fastica
 from .dict_learning import dict_learning, dict_learning_online, sparse_encode,\
                            DictionaryLearning, MiniBatchDictionaryLearning,\
                            SparseCoder
-from .spectra_embedding import SpectralEmbedding
 
 __all__ = ['DictionaryLearning',
            'FastICA',
@@ -26,7 +25,6 @@ __all__ = ['DictionaryLearning',
            'RandomizedPCA',
            'SparseCoder',
            'SparsePCA',
-           'SpectralEmbedding',
            'dict_learning',
            'dict_learning_online',
            'fastica',
