+++ sklearn/manifold/__init__.py
@@ -5,3 +5,4 @@ The :mod:`sklearn.manifold` module implements data embedding techniques.
 from .locally_linear import locally_linear_embedding, LocallyLinearEmbedding
 from .isomap import Isomap
 from .mds import MDS
+from .spectra_embedding import SpectralEmbedding
