+++ sklearn/manifold/spectra_embedding.py
